+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] - 2022-09-28
+
+### Added
+- Added more initial scene load options to `ScriptableSceneController`.
+- Added more guards when loading invalid scenes.
+- 
+### Changed
+- Updated package.json with new version.
+
+### Fixed
+- `ScriptableSceneController` not cleaning up `IsLoading` and `loadingCollection` on exception.
+
 ## [0.0.3] - 2022-09-28
 
 ### Fixed
@@ -25,7 +37,7 @@ Minor UX updates and bug fixes.
 ### Fixed
 - Incorrect collection progress being reported when a collection is loading.
 
-## [0.0.1] - 2022-05-21
+## [v0.0.1] - 2022-05-21
 Initial preview version.
 
 ### Added